Move RepeatView to utils
